+++ src/main/java/kitchenpos/dto/request/order/CreateOrderRequest.java
@@ -1,6 +1,7 @@
 package kitchenpos.dto.request.order;
 
 import java.util.List;
+import kitchenpos.dto.OrderLineItemsDto;
 
 public class CreateOrderRequest {
