+++ src/main/java/kitchenpos/dto/request/table/ChangeEmptyRequest.java
@@ -3,14 +3,7 @@ package kitchenpos.dto.request.table;
 public class ChangeEmptyRequest {
     private Boolean empty;
 
-    public ChangeEmptyRequest() {
-    }
-
     public Boolean getEmpty() {
         return empty;
-    }
-
-    public void setEmpty(Boolean empty) {
-        this.empty = empty;
     }
 }
